fixed blurry text

+++ Caveworks/Scenes/CreditsScreen.cs
@@ -20,7 +20,7 @@ namespace Caveworks.GameScreens
         }
         public static void Draw()
         {
-            Game.mainSpriteBatch.DrawString(Fonts.menuButtonFont, "Made by: David Sobek", new Vector2((GameWindow.windowSize.X - Fonts.menuButtonFont.MeasureString("Made by: David Sobek").X) / 2, GameWindow.windowSize.Y / 2), Color.Yellow);
+            Game.mainSpriteBatch.DrawString(Fonts.menuButtonFont, "Made by: David Sobek", new Vector2((int)((GameWindow.windowSize.X - Fonts.menuButtonFont.MeasureString("Made by: David Sobek").X) / 2), (int)(GameWindow.windowSize.Y / 2)), Color.Yellow);
         }
     }
 }
